Reworked InterceptionModel, added MetadataCachingReader
A previous refactoring has made the InterceptionModel use only ClassMetadata<?> for representing interceptors. This is not working properly in Weld, because interceptors can be Bean references.

Instead of ClassMetadata, we will use InterceptorReference to represent different types of interceptors and to provide various instantiation algorithms. An InterceptorReference is required to return the ClassMetadata<?> associated with the interceptor. An InterceptionModel is built out of InterceptorMetadata.

Additionally, a MetadataCachingReader (thread-safe) is provided for simplifying the metadata reading operations. The client will need to register a singleton instance of it, using a mechanism that is specific to it.